RA&B created an connected with controlAgent

+++ controlAgent/controlAgent.py
@@ -1,20 +1,17 @@
-from flask import Flask, session, request,jsonify
+from flask import Flask, request,jsonify, abort
 from databaseManager import DatabaseManager
-
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.primitives import padding
-from cryptography.hazmat.backends import default_backend
 from cryptoManager import CryptoManager
-import base64
-import hashlib
+from configClasses.rabConfig import RabConfig
 import uuid
 import os 
+import requests
 
 class ControlAgent: 
     def __init__(self):
         self.app = Flask(__name__)    
         self.setup_routes()
         self.DBusers = DatabaseManager()
+        self.rabConfig = RabConfig()
         
         
     def setup_routes(self):
@@ -35,6 +32,7 @@ class ControlAgent:
         # Responder con un mensaje de éxito
         return jsonify({"message": "Datos recibidos con éxito"})
     
+    
     def createNewDatabase(self):
         json = request.get_json()
     
@@ -45,6 +43,10 @@ class ControlAgent:
         elementNames = json.get ('elementNames')
         content = json.get('content')
         
+        result = self.DBusers.verify_user(user, password)
+        if (not result):
+            abort(403, description="Invalid credentials")
+            
         if None in (user, cryptpassword, database, elementNames, content):
             missing_fields = [key for key in ['user', 'pass', 'database', 'elementNames', 'content'] if json.get(key) is None]
             return jsonify({"error": "Faltan argumentos en el JSON", "missing_fields": missing_fields}), 400
@@ -53,35 +55,59 @@ class ControlAgent:
         files = content.split('\n')
         
         folder_name = str(uuid.uuid4())
-        os.makedirs(folder_name, exist_ok=True)
-        save_path = os.path.join ("controlAgent",folder_name,fileNames[0])
+        folder_path = os.path.join("controlAgent", folder_name)
+        os.makedirs(folder_path)
         
-        CryptoManager.decrypt_pdf(files[0], save_path)
+        for i in range(len(files) - 1):
+            save_path = os.path.join ("controlAgent",folder_name,fileNames[i])
+            CryptoManager.decrypt_pdf(files[i], save_path)
             
+        ##Send to db Agent to create Database
+        endpoint = "/createvectordatabase"
+        URL = f"{self.rabConfig.ip}{endpoint}"
+        contentString=""
+        elementNames = ""
+        uploadPath = folder_path
+        
+        for file in os.listdir(uploadPath):
+            filePath = os.path.join(uploadPath, file)
+            if os.path.isfile(filePath) and file.lower().endswith('.pdf'):
+                with open(filePath, 'rb') as file:
+                    auxName = file.name.split(os.sep)
+                    name = auxName[len(auxName) -1]
+                    elementNames = elementNames + name + '#'
+                    encryptedFile = CryptoManager.encrypt_pdf(filePath, self.rabConfig.cypherPass)
+                    contentStringaux = encryptedFile + '\n'
+                    contentString = contentString + contentStringaux
+               
+          
+        data = {
+            "elementNames": elementNames,
+            "content" : contentString
+        }
+
+        response = requests.post(URL, json=data)
+            
+        self.delete_directory(folder_path)    
         return "hi"
 
 
-                
+    def delete_directory(self, path):
+        if os.path.exists(path):            
+            if os.path.isdir(path):
+                for item in os.listdir(path):
+                    item_path = os.path.join(path, item)
+                    if os.path.isdir(item_path):
+                        self.delete_directory(item_path)
+                    else:
+                        os.remove(item_path)
+                os.rmdir(path) 
                 
     def run(self):
         self.app.run(port=5006, debug=True)
         
         
-    def delete_folder_and_contents(path):
-        if os.path.exists(path):
-            # Recursively delete all files and subdirectories
-            for root, dirs, files in os.walk(path, topdown=False):
-                for name in files:
-                    file_path = os.path.join(root, name)
-                    os.remove(file_path)
-                for name in dirs:
-                    dir_path = os.path.join(root, name)
-                    os.rmdir(dir_path)
-            # Remove the main directory
-            os.rmdir(path)
-            print(f"The folder '{path}' and all its contents have been successfully deleted.")
-        else:
-            print(f"The folder '{path}' does not exist.")
+
 
 if __name__ == '__main__':
     controlAgent = ControlAgent()
